feat: use contract registry to resolve addresses
