+++ buidler.config.js
@@ -30,6 +30,6 @@ module.exports = {
     },
   },
   solc: {
-    version: "0.6.8",
+    version: '0.6.8',
   },
 };
