+++ contracts/StakingBank.sol
@@ -1,6 +1,8 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.6.8;
 
+import "@nomiclabs/buidler/console.sol";
+
 import '@openzeppelin/contracts/utils/ReentrancyGuard.sol';
 import '@openzeppelin/contracts/token/ERC20/ERC20.sol';
 import '@openzeppelin/contracts/token/ERC20/SafeERC20.sol';
@@ -9,25 +11,24 @@ import "./interfaces/IERC20Detailed.sol";
 import "./interfaces/IStakingBank.sol";
 import "./interfaces/IValidatorRegistry.sol";
 
-contract StakingBank is IStakingBank, ERC20, ReentrancyGuard {
+import "./extensions/Registrable.sol";
+
+import "./Registry.sol";
+
+contract StakingBank is IStakingBank, ERC20, ReentrancyGuard, Registrable {
   using SafeERC20 for IERC20;
   using SafeMath for uint256;
 
-  IERC20 public token;
-  IValidatorRegistry public registry;
-
-  constructor(
-    address _tokenAddress,
-    address _registryAddress
-  ) ERC20(
-      string(abi.encodePacked("staked ", IERC20Detailed(_tokenAddress).name())),
-      string(abi.encodePacked("sb", IERC20Detailed(_tokenAddress).symbol()))
+  constructor(address _contractRegistry, string memory _name, string memory _symbol)
+  Registrable(_contractRegistry)
+  ERC20(
+      string(abi.encodePacked("staked ", _name)),
+      string(abi.encodePacked("sb", _symbol))
     ) public {
-    require(_tokenAddress != address(0x0), "_tokenAddress is missing");
-    require(_registryAddress != address(0x0), "_registryAddress is missing");
+  }
 
-    token = IERC20(_tokenAddress);
-    registry = IValidatorRegistry(_registryAddress);
+  function getName() override external pure returns (bytes32) {
+    return 'StakingBank';
   }
 
   function receiveApproval(
@@ -35,6 +36,9 @@ contract StakingBank is IStakingBank, ERC20, ReentrancyGuard {
     uint256 _value,
     bytes calldata _data
   ) override external nonReentrant returns (bool success) {
+    IValidatorRegistry registry = validatorRegistryContract();
+    IERC20 token = tokenContract();
+
     (address id, ) = registry.validators(_from);
 
     require(id != address(0x0), "validator does not exist in registry");
@@ -49,6 +53,7 @@ contract StakingBank is IStakingBank, ERC20, ReentrancyGuard {
   }
 
   function withdraw(uint256 _value) override external nonReentrant returns (bool success) {
+    IERC20 token = tokenContract();
     uint256 balance = balanceOf(msg.sender);
 
     require(_value > 0, "empty withdraw value");
