+++ example.env
@@ -1,6 +1,7 @@
 STAGING_PK=
+VALIDATOR_PK=
 INFURA_ID=
+
+# below are just for running local scripts like block-listener
 LOCAL_BLOCK_TIME=15
 CHAIN_ADDRESS=
-VALIDATOR_PK=
-OWNER_PK=
