+++ package.json
@@ -12,11 +12,12 @@
     "test": "buidler test",
     "local-minter": "npx buidler run --network localhost ./scripts/minter.js",
     "block-listener": "npx buidler run --network localhost scripts/block-listener.js",
-    "deploy:registry:local": "npx buidler run --network localhost ./scripts/deploy-registry.js",
-    "deploy:registry:staging": "NODE_ENV=staging npx buidler run --network kovan ./scripts/deploy-registry.js",
-    "deploy-poc:local": "npx buidler run --network localhost ./scripts/deploy.js",
-    "deploy-poc:staging": "NODE_ENV=staging npx buidler run --network kovan ./scripts/deploy.js",
-    "deploy-poc": "npx buidler run ./scripts/deploy.js",
+    "deploy:registry": "npx buidler run --network localhost scripts/deploy-registry.js",
+    "deploy:registry:staging": "NODE_ENV=staging npx buidler run --network kovan scripts/deploy-registry.js",
+    "deploy:contracts": "npx buidler run --network localhost scripts/deploy-contracts.js",
+    "deploy:contracts:staging": "NODE_ENV=staging npx buidler run --network kovan scripts/deploy-contracts.js",
+    "deploy:all": "npx buidler run --network localhost scripts/deploy-all.js",
+    "deploy:all:staging": "NODE_ENV=staging npx buidler run --network kovan scripts/deploy-all.js",
     "flatten:token": "npx truffle-flattener ./contracts/Token.sol | awk '/SPDX-License-Identifier/&&c++>0 {next} 1' > _token.sol",
     "flatten:validator": "npx truffle-flattener ./contracts/ValidatorRegistry.sol | awk '/SPDX-License-Identifier/&&c++>0 {next} 1' > _validatorRegistry.sol",
     "flatten:chain": "npx truffle-flattener ./contracts/Chain.sol | awk '/SPDX-License-Identifier/&&c++>0 {next} 1' > _chain.sol",
